+++ packages/common/src/createAccount.ts
@@ -30,6 +30,7 @@ export const createAccount = (config: Config) =>
         },
       },
     }),
+    debugName: 'createAccount',
   })
     .then((business: any) => ({
       ...config,