+++ packages/platform/src/Metadata/index.ts
@@ -1,6 +1,5 @@
-export { Inject, Injectable, ReflectiveInjector } from 'injection-js';
-
 export { CloudServiceInstance } from './CloudServiceInstance';
 export * from './Context';
+export * from '@zetapush/common/lib/metadata';
 export { Options } from './Options';
 export * from './Bootstrapable';