+++ packages/worker/rollup.config.js
@@ -2,7 +2,7 @@ import resolve from 'rollup-plugin-node-resolve';
 import commonjs from 'rollup-plugin-commonjs';
 import json from 'rollup-plugin-json';
 import ts from 'rollup-plugin-typescript';
-import typescript from 'typescript'
+import typescript from 'typescript';
 import { uglify } from 'rollup-plugin-uglify';
 import { minify as minifier } from 'uglify-es';
 
@@ -12,12 +12,20 @@ const config = {
     resolve({
       jsnext: true,
       main: true,
+      // browser: true,
     }),
     commonjs({
-      include: ['node_modules/**', '../client/**', '../cometd/**', '../core/**', '../platform/**']
+      include: [
+        'node_modules/**',
+        '../client/**',
+        '../cometd/**',
+        '../core/**',
+        '../platform/**',
+        '../common/**',
+      ],
     }),
     ts({
-      typescript
+      typescript,
     }),
     json(),
   ],
@@ -25,23 +33,27 @@ const config = {
     format: 'umd',
     globals: {
       '@zetapush/client': 'ZetaPushClient',
-      '@zetapush/platform': 'ZetaPushPlatform'
+      '@zetapush/common': 'ZetaPushCommon',
+      '@zetapush/platform': 'ZetaPushPlatform',
     },
     name: 'ZetaPushWorker',
-    sourcemap: true
+    sourcemap: true,
   },
 };
 
 if (env === 'production') {
   config.plugins.push(
-    uglify({
-      compress: {
-        pure_getters: true,
-        unsafe: true,
-        unsafe_comps: true,
-        warnings: false,
+    uglify(
+      {
+        compress: {
+          pure_getters: true,
+          unsafe: true,
+          unsafe_comps: true,
+          warnings: false,
+        },
       },
-    }, minifier),
+      minifier,
+    ),
   );
 }