+++ packages/worker/src/index.ts
@@ -1,4 +1,2 @@
 export { WorkerClient, Worker } from './client/worker';
 export * from './client/run';
-export * from './utils/compress';
-export * from './utils/upload';